Move functional tests to a separate file
